+++ CHANGELOG.md
@@ -184,6 +184,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1592](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1592))
 - `opentelemetry-instrumentation-django` Allow explicit `excluded_urls` configuration through `instrument()`
   ([#1618](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1618))
+- Added query sanitization for pymemcache
+  ([#1576](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1576))
 
 ### Fixed